Update code QMM002-C

+++ nts.uk/ctx/basic/nts.uk.ctx.basic.app/src/main/java/nts/uk/ctx/basic/app/command/system/bank/branch/RemoveBranchCommandHandler.java
@@ -1,9 +1,13 @@
 package nts.uk.ctx.basic.app.command.system.bank.branch;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 
+import nts.arc.error.BusinessException;
 import nts.arc.layer.app.command.CommandHandler;
 import nts.arc.layer.app.command.CommandHandlerContext;
 import nts.uk.ctx.basic.dom.system.bank.branch.BankBranchRepository;
@@ -23,7 +27,13 @@ public class RemoveBranchCommandHandler extends CommandHandler<RemoveBranchComma
 	protected void handle(CommandHandlerContext<RemoveBranchCommand> context) {
 		String companyCode = AppContexts.user().companyCode();
 		RemoveBranchCommand command = context.getCommand();
-		
+		List<String>  bankCodeList = new ArrayList<String>();
+	    bankCodeList.add(command.getBankCode());
+	    List<String> branchCodeList = new ArrayList<String>();
+	    branchCodeList.add(command.getBranchCode());
+	    if(personBankAccountRepository.checkExistsBranchAccount(companyCode, bankCodeList, branchCodeList)){
+	    	throw new BusinessException("選択された＊は使用されているため削除できません。"); // ER008
+	    }
 		banhBranchRepo.remove(companyCode, command.getBankCode(), command.getBranchCode());	
 	}
